requirements.txt
fastapi 
uvicorn
jinja2
google-auth-httplib2
google-api-python-client
pydantic
python-multipart